Setup pytest and test first endpoint
The _create classmethod of the DjangoModelFactory used in the ExpenseFactory
required overriding for snapshot stability of the datetime values.
The auto_now_add and auto_now attributes of the DateTimeFields on the Expense
model were varying with every snapshot as the time that the test was
ran changed.
The override removes the problematic field attributes only when the
Factory is used.